Adds project saving functionality
Implements the ability to save project data to a JSON file.

This change introduces `encode` methods to the `Module` and `Project` classes, allowing objects to be serialized into JSON format. It also adds a "Save Project" action to the CocomoApp, which writes the project data to a `project.json` file within a `projects` directory.

Adds project directory to gitignore.