Adds save screen for project persistence
Introduces a modal screen for specifying the save path and filename when saving a project.

This screen allows the user to choose the location and name of the JSON file where the project data will be stored.

Also implements the callback to handle the result of the save screen, persisting the project data to the specified file upon confirmation, creating the parent directories if necessary.